% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNodeEdgeFiles.R
\name{BuildEdgeFile}
\alias{BuildEdgeFile}
\title{Builds a final edge datafile from the set of relevant data files}
\usage{
BuildEdgeFile(path_to_graph, path_to_edge_weights, showProgress = F)
}
\arguments{
\item{path_to_graph}{Path to the graph file contianing the node labels and
coordiantes}

\item{path_to_edge_weights}{Path to the .csv file contianing edge weights}

\item{showProgress}{Boolean, whether to show progress when loading using
the fread function.}

\item{outfile}{Path to save output edge data file}
}
\description{
Builds a final edge datafile from the set of relevant data files
}
