% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNetwork.R
\name{BuildNetwork}
\alias{BuildNetwork}
\title{Plot a network in NSP style given NSP node and edge data}
\usage{
BuildNetwork(
  nodes,
  edges,
  node.color = "black",
  node.colors = NULL,
  node.color.levels = NULL,
  node.color.name = "Discipline",
  node.size = 5,
  node.sizes = NULL,
  max.node.size = 10,
  node.size.breaks = NULL,
  node.size.labels = NULL,
  node.size.name = "Publication Count",
  node.show = NULL,
  node.stroke = 1.1,
  node.alpha = 1,
  edge.color = "grey",
  edge.size = 0.2,
  edge.alpha = 0.4,
  viridis.option = "C",
  show.legend.size = TRUE,
  show.legend.color = TRUE
)
}
\arguments{
\item{nodes}{A dataframe containing the Specialty/Discipline labels of nodes,
along with corresponding x and y coordinates. The node labels should be
stored in a dataframe column titled "Specialty".}

\item{edges}{A dataframe containing the edges data for the network, in the
format returned from the igraph function \code{as_data_frame(g, what = "edges")}}

\item{node.color}{Color of all nodes, if no node-wise colors are defined}

\item{node.colors}{Vector containing factor values dictating colors of nodes,
in same order as nodes appear in the `nodes` dataframe}

\item{node.color.levels}{Factor levels for the `node.colors``variable}

\item{node.color.name}{The title of the color legend}

\item{node.size}{Size of all nodes, if no node-wise sizes are provided}

\item{node.sizes}{Vector containing sizes of all nodes, in same order as they
appesr in the `nodes` dataframe}

\item{max.node.size}{Maximum size, in regards to area, of nodes when plotting}

\item{node.size.breaks}{A vector of breaks to be used in the node size legend.}

\item{node.size.labels}{A vector of labels to be used in the node size legend.
Must be the same size as the `node.size.breaks` variable}

\item{node.size.name}{The title of the size legend}

\item{node.show}{A logical vector specifying whether each node should be made
visible, of the same size and orders of the `nodes` dataframe}

\item{node.stroke}{The size of the border of the nodes}

\item{node.alpha}{The alpha value of visible nodes}

\item{edge.color}{The color of edges}

\item{edge.size}{The size of edges}

\item{viridis.option}{For plotting with viridis colors. "A character string
indicating the colormap option to use. Four options are available: "magma"
(or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the
default option) and "cividis" (or "E")."}

\item{show.legend.size}{Boolean value, whether or not to show color legend}
}
\value{
A ggplot2 object containing the NSP style network visualization
}
\description{
Plot a network in NSP style given NSP node and edge data
}
\examples{
nodes <- get_node_data()
edges <- get_edge_data()

usa_nodes <- get_country_year_nodes(nodes)

# Plot network with defaul options
build_nsp_network(usa_nodes, edges)

# Plot network with custom colors
build_nsp_network(usa_nodes, edges, node.colors = usa_nodes$level_3, node.size = 3)
}
