% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildNodeEdgeFiles.R
\name{BuildNodeFile}
\alias{BuildNodeFile}
\title{Builds the node and edge datafiles from raw datafiles}
\usage{
BuildNodeFile(
  path_to_graph,
  path_to_edge_weights,
  path_to_count,
  path_to_rca,
  path_to_disciplines,
  mirror.nodes.x = FALSE,
  mirror.nodes.y = FALSE,
  showProgress = F
)
}
\arguments{
\item{path_to_graph}{Path to the graph file contianing the node labels and
coordiantes}

\item{path_to_count}{Path the the .csv file contianing the bibliometric counts}

\item{path_to_rca}{Path to the .csv file contianing the calculated RCA values}

\item{path_to_disciplines}{Path to the .csv file conitaning the disciplinary
classification hierarchy}

\item{showProgress}{Boolean, whether to show progress when loading using
the fread function.}

\item{outfile}{Path to save output node data file}
}
\description{
Builds the node and edge datafiles from raw datafiles
}
