% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildWorldMap.R
\name{BuildWorldMap}
\alias{BuildWorldMap}
\title{Plot the RCA world maps in NSP-style}
\usage{
BuildWorldMap(
  nodes,
  discipline,
  years = NULL,
  mapping = NULL,
  high.rca.alpha = 1,
  low.rca.alpha = 0.3,
  min.log.rca = -2,
  max.log.rca = 4,
  plot.viridis = F,
  viridis.option = "C",
  show.color.legend = TRUE,
  rca.threshold = FALSE,
  legend.title = "Log Revealed Comparative Advantage"
)
}
\arguments{
\item{nodes}{A dataframe containing the Specialty/Discipline labels of
nodes, along with corresponding bibliometric data. The node labels
should be stored in a dataframe column titled "Specialty".}

\item{discipline}{A string mapping to a discipline to represent on
the world map}

\item{years}{A string mapping to a year interval to use to filter
RCA values. Defaulta to the most recent time period}

\item{mapping}{data file containing mapping between the country names
used in the Web of Science, and those expected by ggplot. Expected to
have two columns: `region` and `Mapping`, where `region` is the
ggplot name, and `Mapping` is the web of science name.}

\item{high.rca.alpha}{The alpha value for countries with high RCA}

\item{low.rca.alpha}{The alpha value for countries with low RCA}

\item{plot.viridis}{Boolean, whether or not to plot with viridis. Requires
that the viridis package is installed}

\item{viridis.option}{For plotting with viridis colors. "A character string
indicating the colormap option to use. Four options are available: "magma"
(or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the
default option) and "cividis" (or "E")."}

\item{show.color.legend}{Boolean, whether or not to show the color legend}
}
\description{
Plot the RCA world maps in NSP-style
}
\examples{

# Use default parameters
build_nsp_worldmap(nodes, discipline = "Sociology")

# Plot for older interval
build_nsp_worldmap(nodes, discipline = "Sociology", years = "2008-2011")

}
