% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterNodes.R
\name{FilterNodes}
\alias{FilterNodes}
\title{Filters the NSP node dataframe to that of a select country or year, or,
aggregates data over a range of conutries and years}
\usage{
FilterNodes(node_table, country = NULL, years = NULL, showWarnings = TRUE)
}
\arguments{
\item{node_table}{The node datafile to filter or aggregate, in NSP long
format}

\item{country}{A select country or vector of countries to filter by or
aggregate over}

\item{years}{A select year interval to vector of intervals to filter by
or aggreegate over. All intervals must also appear in the in the}

\item{showWarnings}{Boolean, whetehr to show warnings.}
}
\description{
In the case that a list of countries or years are provided,
the RCA is calculated as a mean of all the RCA values of the countries,
weighted by the total number of papers.
}
\examples{
# aggregates over all countries and years
filter_nsp_nodes(nodes)

# aggregates USA data over all years
filter_nsp_nodes(nodes, country = "United States")

# Aggregate over all countries for a specified time interval
filter_nsp_data(nodes, years = "2012-2016")

# Filter to a single country and year
filter_nsp_data(nodes, country = "United States", years = "2012-2016")

# Filter to a list of countries and a list of years
filter_nsp_data(nodes, years = c("2008-2011", "2012-2016"), conutry = c("Germany", "France"))

}
