% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFile.R
\name{LoadFile}
\alias{LoadFile}
\title{Load a single RCA or proximity file. These files were orignially
saved as a .csv file contianing a matrix, with a row for each
country and a column for each discipline}
\usage{
LoadFile(path.to.file, type = NULL)
}
\arguments{
\item{path.to.file}{Path to the RCA or proximity file to load. The file
should be formatted as a matrix, such that the first column are the names
of countries, and each column is the name of a discipline}

\item{type}{A string representing whether the file contains an RCA, or a
Proximity data}
}
\value{
A dataframe containing the RCA/proximity meausre mapped to each
country, disicpline, and year interval
}
\description{
Load a single RCA or proximity file. These files were orignially
saved as a .csv file contianing a matrix, with a row for each
country and a column for each discipline
}
