% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTernary.R
\name{PlotTernary}
\alias{PlotTernary}
\title{Plot a network in NSP style given NSP node and edge data}
\usage{
PlotTernary(
  data,
  n = 100,
  point.size = 0.3,
  plot.legend = FALSE,
  gradient.max = NA,
  text.size = 30,
  title.size = 30
)
}
\description{
Plot a network in NSP style given NSP node and edge data
}
