% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceYearWithYears.R
\name{ReplaceYearWithYears}
\alias{ReplaceYearWithYears}
\title{Builds the node and edge datafiles from raw datafiles}
\usage{
ReplaceYearWithYears(year)
}
\arguments{
\item{year}{the single year value to replace with NSP time period equivelant}
}
\description{
Builds the node and edge datafiles from raw datafiles
}
